Consistently declare block incremental size as size_t.
The block is often allocated in memory so size_t makes more sense than uint64_t.

+++ src/command/backup/file.c
@@ -250,8 +250,7 @@ backupFile(
                         ioFilterGroupAdd(
                             ioReadFilterGroup(storageReadIo(read)),
                             blockIncrNew(
-                                (size_t)file->blockIncrSize, blockIncrReference, bundleId, bundleOffset, blockMap, compress,
-                                encrypt));
+                                file->blockIncrSize, blockIncrReference, bundleId, bundleOffset, blockMap, compress, encrypt));
 
                         repoChecksum = true;
                     }
